Refactor number sequence unescaping
